refactor: move db logic outside of the models.py into the db.py
